CPCSS admin classes architecture update (PR #2687)
* Add Post & Settings classes
* Move & rename subscriber, update dependencies
* update classes instantiation
* update all tests
