+++ tests/Fixtures/inc/Engine/CriticalPath/Admin/Admin/enqueueAdminCpcssHeartbeatScript.php
@@ -1,0 +1,20 @@
+<?php
+
+return [
+	'testShouldNotEnqueueScript' => [
+		'config'   => [
+			'options' => [
+				'async_css' => 0,
+			],
+		],
+		'expected' => false,
+	],
+	'testShouldEnqueueScript' => [
+		'config'   => [
+			'options' => [
+				'async_css' => 1,
+			],
+		],
+		'expected' => true,
+	],
+];
