+++ tests/Integration/inc/Engine/CriticalPath/Admin/Subscriber/setAsyncCssMobileDefaultValue.php
@@ -1,21 +1,38 @@
 <?php
 
-namespace WP_Rocket\Tests\Integration\inc\Engine\CriticalPath\AdminSubscriber;
+namespace WP_Rocket\Tests\Integration\inc\Engine\CriticalPath\Admin\Subscriber;
 
 use WP_Rocket\Tests\Integration\TestCase;
 
 /**
- * @covers \WP_Rocket\Engine\CriticalPath\AdminSubscriber::set_async_css_mobile_default_value
+ * @covers \WP_Rocket\Engine\CriticalPath\Admin\Subscriber::set_async_css_mobile_default_value
  *
  * @group  AdminOnly
  * @group  CriticalPath
+ * @group  CriticalPathAdminSubscriber
  */
 class Test_SetAsyncCssMobileDefaultValue extends TestCase {
+    use ProviderTrait;
+
+    protected static $class_name = 'Settings';
+
+    public function setUp() {
+        parent::setUp();
+
+        remove_action( 'wp_rocket_upgrade', 'rocket_new_upgrade' ); 
+    }
+
+    public function tearDown() {
+        parent::tearDown();
+
+        add_action( 'wp_rocket_upgrade', 'rocket_new_upgrade' ); 
+    }
 
 	/**
-	 * @dataProvider configTestData
+	 * @dataProvider providerTestData
 	 */
 	public function testShouldUpdateOption( $versions, $update ) {
+        
         do_action( 'wp_rocket_upgrade', $versions['new'], $versions['old'] );
 
         $options = get_option( 'wp_rocket_settings', [] );
@@ -36,5 +53,5 @@ class Test_SetAsyncCssMobileDefaultValue extends TestCase {
                 $options
             );
         }
-	}
+    }
 }
