+++ tests/Unit/inc/Engine/CriticalPath/Admin/Settings/displayCpcssMobileSection.php
@@ -1,9 +1,12 @@
 <?php
 
-namespace WP_Rocket\Tests\Unit\inc\Engine\CriticalPath\AdminSubscriber;
+namespace WP_Rocket\Tests\Unit\inc\Engine\CriticalPath\Admin\Settings;
 
 use Brain\Monkey\Functions;
+use Mockery;
 use WP_Rocket\Tests\Unit\TestCase;
+use WP_Rocket\Engine\CriticalPath\Admin\Settings;
+use WP_Rocket\Tests\Unit\inc\Engine\CriticalPath\Admin\AdminTrait;
 
 /**
  * @covers \WP_Rocket\Engine\CriticalPath\AdminSubscriber::cpcss_section
@@ -11,26 +14,30 @@ use WP_Rocket\Tests\Unit\TestCase;
  * @group  CriticalPath
  */
 class Test_DisplayCpcssMobileSection extends TestCase {
-	use GenerateTrait;
+	use AdminTrait;
 
 	protected static $mockCommonWpFunctionsInSetUp = true;
+
+	private $settings;
 
 	protected function setUp() {
 		parent::setUp();
 
 		$this->setUpMocks();
-	}
 
-	protected function tearDown() {
-		parent::tearDown();
+		$this->settings = Mockery::mock( Settings::class . '[generate]', [
+				$this->options,
+				$this->beacon,
+				$this->critical_css,
+				WP_ROCKET_PLUGIN_ROOT . 'views/cpcss/',
+			]
+		);
 	}
 
 	/**
 	 * @dataProvider configTestData
 	 */
 	public function testShouldDisplayCPCSSMobileSection( $config, $expected ) {
-		$this->setUpMocks();
-
 		foreach ( $config['options'] as $option_key => $option ) {
 			$this->options
 				->shouldReceive( 'get' )
@@ -48,10 +55,12 @@ class Test_DisplayCpcssMobileSection extends TestCase {
 
 		Functions\when( 'current_user_can' )->justReturn( $config['current_user_can'] );
 
-		$this->setUpGenerate( 'activate-cpcss-mobile', ['beacon' => $config['beacon'] ] );
+		$this->settings->shouldReceive( 'generate' )
+				   ->with( 'activate-cpcss-mobile', ['beacon' => $config['beacon'] ] )
+				   ->andReturn( '' );
 
 		ob_start();
-		$this->subscriber->display_cpcss_mobile_section();
+		$this->settings->display_cpcss_mobile_section();
 		ob_get_clean();
 
 	}
